requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.2
django-extensions==3.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
drf-nested==1.2.22
nested-multipart-parser==1.3.1
psycopg2-binary==2.9.4
PyJWT==2.5.0
pytz==2022.4
sqlparse==0.4.3

python-dateutil~=2.8.2
pandas~=1.5.1